requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data sources
yfinance==0.2.33
arxiv==2.1.0

# Scientific computing
scipy==1.11.4
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard
streamlit==1.29.0

# Configuration
pyyaml==6.0.1
tomli==2.0.1

# Progress bars
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
